Renamed to client sub module
