+++ tests/unit_tests/providers/opensearch_test.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from sciety_labs.providers.opensearch.opensearch import (
+from sciety_labs.providers.opensearch.client import (
     OpenSearchConnectionConfig,
     OpenSearchEnvVariables
 )
